+++ AdaTranslation.Domain/Entities/UserLanguage.cs
@@ -3,7 +3,7 @@
     public class UserLanguage
     {
         public int Id { get; private set; }
-        public int UserId { get; private set; }
+        public long UserId { get; private set; }
         public int LanguageId { get; private set; }
 
         public User User { get; private set; } = null!;
